% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birt.R
\name{add_fit_indices}
\alias{add_fit_indices}
\title{Compute and store the marginal-likelihood fit indices}
\usage{
add_fit_indices(fit, n_draws = 1000, Q = 41, progress = TRUE)
}
\arguments{
\item{fit}{A `birt` object.}

\item{n_draws}{Number of posterior draws used (reduced automatically when ALD adds
quadrature layers).}

\item{Q}{Gauss-Hermite nodes.}

\item{progress}{Show progress.}
}
\value{
`fit` with `fit_indices`, `loglik`, `loo` and `waic` added.
}
\description{
Integrates the latent variables out (see Details of [birt()]) to get the pointwise
log-likelihood, then DIC, WAIC and PSIS-LOO. Called by [birt()] unless
`fit_indices = FALSE`.
}
