% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_birt_fit}
\alias{assert_birt_fit}
\title{Check if object is a birt model fit (internal)}
\usage{
assert_birt_fit(object)
}
\arguments{
\item{object}{Any object.}
}
\value{
TRUE if it's a birt fit, otherwise throws an error.
}
\description{
Check if object is a birt model fit (internal)
}
\keyword{internal}
