% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birt-package.R
\docType{package}
\name{birt-package}
\alias{birt-package}
\title{birt: BIvariate Response Time Modeling with 'lavaan'-Style Syntax}
\description{
Main function [birt()]: by default RTMB with ELGM (approximate Bayes on the marginal likelihood);
maximum likelihood with `rtmb_control(method = "aghq")`; the optional JAGS engine with
`engine = "jags"`. Syntax helper [rtirt_syntax()];
residual distributions [ald()], [normal()], [shash()]; results via `summary()`,
[estimates()], [scores()], [reliability()], [fit_indices()],
[quantiles()], [ppc()], [convergence()], [compare()], `plot()`; JAGS code via [jags_code()].
See the tutorial (`system.file("tutorial", "birt_tutorial.qmd", package = "birt")`).
}
\details{
The methods of the generics shared with birtRcpp (`estimates`, `scores`, `reliability`,
`fit_indices`, `convergence`) are also exported, so they dispatch whichever package was
attached last.
}
\author{
\strong{Maintainer}: Jie-Wen Tsai \email{tsai.jiewen@gmail.com}

}
\keyword{internal}
