% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{birt_quantile}
\alias{birt_quantile}
\alias{estimates.birt_quantile}
\alias{coef.birt_quantile}
\alias{reliability.birt_quantile}
\alias{fit_indices.birt_quantile}
\alias{print.birt_quantile}
\alias{summary.birt_quantile}
\alias{plot.birt_quantile}
\title{Quantile fits: several quantiles of one latent variable or indicator}
\usage{
\method{estimates}{birt_quantile}(object, ...)

\method{coef}{birt_quantile}(object, ...)

\method{reliability}{birt_quantile}(object, ...)

\method{fit_indices}{birt_quantile}(object, ...)

\method{print}{birt_quantile}(x, digits = 3, ...)

\method{summary}{birt_quantile}(object, ...)

\method{plot}{birt_quantile}(x, terms = NULL, ...)
}
\arguments{
\item{object, x}{A `birt_quantile` object.}

\item{...}{Unused.}

\item{digits}{Number of decimals.}

\item{terms}{Regression terms to plot (default: all regressions of the target).}
}
\description{
`birt(..., family = list(speed = ald(c(.1, .5, .9))))` fits one model per quantile
and returns a `birt_quantile` object: `fits` (list of `birt` objects), `p`, `target`.
}
