% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_name}
\alias{get_model_name}
\title{Get model name (internal)}
\usage{
get_model_name(object)
}
\arguments{
\item{object}{A birt fit object.}
}
\value{
Character string: "Rasch", "2PL", or "3PL".
}
\description{
Get model name (internal)
}
\keyword{internal}
