% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{item_fit}
\alias{item_fit}
\title{Item Fit Statistics (Outfit and Infit)}
\usage{
item_fit(object)
}
\arguments{
\item{object}{A birt_fit object.}
}
\value{
A data frame with columns: item, obs_prop, exp_prop, outfit, infit.
}
\description{
Outfit = mean of standardized squared residuals (sensitive to outliers).
Infit  = information-weighted mean-square (sensitive to systematic misfit).
}
\details{
Both should be near 1.0:
Values greater than 1.3 means more noise than the model expects
Values less than 0.7 means less noise than expected (items too predictable)
}
