% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgrm.R
\name{ordinal_information}
\alias{ordinal_information}
\title{Test information of the ordinal indicators}
\usage{
ordinal_information(object, at = NULL, theta = seq(-3, 3, by = 0.1))
}
\arguments{
\item{object}{A [birt()] fit with ordinal indicators (either engine).}

\item{at}{Named list of moderator values (default: the 10th, 50th and 90th percentile of
the first moderator, the others at their mean).}

\item{theta}{Grid of values of the latent variable.}
}
\value{
Data frame with `theta`, the moderator values and the test information
  (point estimates of the item parameters).
}
\description{
For fits with ordinal indicators (graded response or partial credit), the test information of those
indicators as a function of the latent variable, at chosen values of the moderators.
}
