% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_icc}
\alias{plot_icc}
\title{Plot Item Characteristic Curves (ICCs)}
\usage{
plot_icc(object, items = NULL, theta_range = c(-4, 4), ci = TRUE, ...)
}
\arguments{
\item{object}{A birt_fit object.}

\item{items}{Which items to plot (integer vector). Default: all.}

\item{theta_range}{Range of ability values to show. Default c(-4, 4).}

\item{ci}{Show 95\% credible bands? Default TRUE.}

\item{...}{Ignored.}
}
\value{
A ggplot2 object.
}
\description{
An ICC shows the probability of getting an item correct
as a function of ability. Each item gets one curve.
Harder items have curves shifted to the right.
}
