% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_info}
\alias{plot_info}
\title{Test Information Function}
\usage{
plot_info(
  object,
  items = NULL,
  theta_range = c(-4, 4),
  show_items = FALSE,
  ...
)
}
\arguments{
\item{object}{A bayesRasch_fit object.}

\item{items}{Which items to include. Default: all.}

\item{theta_range}{Range of ability values. Default c(-4, 4).}

\item{show_items}{Show individual item curves? Default FALSE.}

\item{...}{Ignored.}
}
\value{
A ggplot2 object.
}
\description{
Shows where on the ability scale the test is most precise.
In Rasch, item info = P(1-P), which peaks at ability = difficulty.
Test info = sum of all item infos.
}
