% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_wright_map}
\alias{plot_wright_map}
\title{Wright Map (Item-Person Map)}
\usage{
plot_wright_map(object, ...)
}
\arguments{
\item{object}{A bayesRasch_fit object.}

\item{...}{Ignored.}
}
\value{
A ggplot2 object.
}
\description{
Shows person abilities (histogram) and item difficulties (triangles)
on the same logit scale. Lets you see if items cover the ability range.
}
