% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.birt_fit}
\alias{print.birt_fit}
\title{Print a Rasch Model Fit}
\usage{
\method{print}{birt_fit}(x, ...)
}
\arguments{
\item{x}{A birt_fit object.}

\item{...}{Ignored.}
}
\description{
Shows a quick overview when you type the fitted object name.
}
