% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{algebra}
\alias{algebra}
\title{Student Assessment Dataset}
\format{
A matrix with 1302 rows (students) and 12 columns (items).

\describe{
\item{Rows}{Students}
\item{Columns}{Test Items}
}
}
\source{
Example assessment dataset included for demonstration.
}
\usage{
algebra
}
\description{
A binary response matrix used to demonstrate the Bayesian Rasch model.
}
\keyword{datasets}
